requirements.txt
discord.py==2.3.2
pandas==2.0.3
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
requests==2.31.0
python-dotenv==1.0.0